Push down recommendation Redirect, 1st render code
